requirements.txt
celery==5.3.1
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
python-decouple==3.8
redis==5.0.0
